add manage profile UI pages
We need the capabilities to add and delete profiles via the UI. This
also resolves the bug that was appearing when trying to create a new
profile in the UI previously.

Bug was due to passing the `container` as the profile to select instead
of the `profile`, because I'm very smart.